requirements.txt
flask
pandas
numpy
open-clip-torch
torch
scikit-learn
pillow